refactored summary components into separate files
